Improve configuration
Before it was kind of clunky to change anything about the database, and
there always needed to be a file in the repo with the settings. Now it
is possible to override the default settings by doing `export
DATABASE_CONFIG=/path/to/config.py`. This change also removes the need
for a temporary file containing the PID of the emailer.

Additionally, it should now be easy to have separate configurations when
testing, debugging, and so on. These settings can be added to the config
folder.